+++ docs/conf.py
@@ -79,7 +79,7 @@ pygments_style = "sphinx"
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
 html_css_files = ["pyscenedetect.css"]
-html_favicon = "favicon.ico"
+html_favicon = "_static/favicon.ico"
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.